requirements.txt
streamlit
openai
weaviate-client
pandas
